Add node for managing game process
Added GameProcess. This manages step-based execution of the game loop.

Game can be set to both human and agent-controlled modes.

Runs the game for a fixed number of physics steps, then pauses and emits a signal to await a new action from the python agent through the TCP server.

Pausing after n steps allows for consistent training when running the game engine at variable speeds.

Provides reset_and_pause, step, and get_observation to control the agent and the rest of the game environment.