Create TCP server class for communicating with python
The TCP server runs on a separate thread and is responsible for connecting with the Python TCP client and getting new actions from the python RL agent.

When receiving training packets from python, it calls methods running on the main thread. These delegate interaction with the game environment to the GameProcess node.