Add main scene with MVP for testing
Main scene containing a full setup for RL training.
This exports variables for setting engine speed, selecting human_mode and choosing if the game should run without rendering (for fast AI training).

The main script starts the process and the TCP server and game process controls the rest after this.