Add scene component to toggle rendering
Disabling rendering hugely reduces the bottleneck from Godot. Without rendering only the physics process and process need to run. This enables the player to toggle rendering while the game is running. This makes it possible to train at maximum speed and also peek at the current progress with rendering enabled.