Change car mechanics
Car is now slower, turns wheel faster and has a lower turn radius.

Removed requirement for car to be stationary to count as parked. This was too difficult and rare for agents to discover. Instead rewards should be shaped to encourage slow parking speeds.

Made collisions with the ground not count as a collision to the agent. Previous agents struggled with fast turns accidentally causing the body to collide with the ground, ending the episode.