Reset parking env obs space limits
Reset car position observations space bounds to previous values.

The trained models use environments with +-np.inf as bounds for car positions. This was changed to 10000 in a refactor. Sb3 requires models to use the same env config for observation spaces to continue training, so these had to be reset.